added some textures, game more smooth

+++ Kaboomer/framebuffer.h
@@ -6,11 +6,11 @@
 
 class FrameBuffer {
 public:
-    size_t w, h; // image dimensions
+    short w, h; // image dimensions
     std::vector<uint32_t> img; // storage container
 
     FrameBuffer();
     void clear(const uint32_t color);
-    void set_pixel(const size_t x, const size_t y, const uint32_t color);
-    void draw_rectangle(const size_t x, const size_t y, const size_t w, const size_t h, const uint32_t color);
+    void set_pixel(const short x, const short y, const uint32_t color);
+    void draw_rectangle(const short x, const short y, const short w, const short h, const uint32_t color);
 };
